Closes #141

+++ src/core/constraint.jl
@@ -12,12 +12,12 @@
 Adds a constraint that ensures flow conservation at a node in the network.
 """
 function constraint_flow_conservation(
-    wm::AbstractWaterModel, n::Int, i::Int, pipe_fr::Array{Int64,1},
-    pipe_to::Array{Int64,1}, des_pipe_fr::Array{Int64,1}, des_pipe_to::Array{Int64,1},
-    pump_fr::Array{Int64,1}, pump_to::Array{Int64,1}, regulator_fr::Array{Int64,1},
-    regulator_to::Array{Int64,1}, short_pipe_fr::Array{Int64,1},
-    short_pipe_to::Array{Int64,1}, valve_fr::Array{Int64,1}, valve_to::Array{Int64,1},
-    reservoirs::Array{Int64,1}, tanks::Array{Int64,1}, dispatchable_demands::Array{Int64,1},
+    wm::AbstractWaterModel, n::Int, i::Int, pipe_fr::Array{Int,1},
+    pipe_to::Array{Int,1}, des_pipe_fr::Array{Int,1}, des_pipe_to::Array{Int,1},
+    pump_fr::Array{Int,1}, pump_to::Array{Int,1}, regulator_fr::Array{Int,1},
+    regulator_to::Array{Int,1}, short_pipe_fr::Array{Int,1},
+    short_pipe_to::Array{Int,1}, valve_fr::Array{Int,1}, valve_to::Array{Int,1},
+    reservoirs::Array{Int,1}, tanks::Array{Int,1}, dispatchable_demands::Array{Int,1},
     fixed_demand::Float64)
     # Collect flow variable references per component.
     q_pipe, q_des_pipe = var(wm, n, :q_pipe), var(wm, n, :q_des_pipe)
@@ -57,7 +57,7 @@ function constraint_tank_volume_fixed(wm::AbstractWaterModel, n::Int, i::Int, V_
 end
 
 
-function constraint_des_pipe_selection(wm::AbstractWaterModel, n::Int, k::Int, node_fr::Int, node_to::Int, des_pipes::Array{Int64,1})
+function constraint_des_pipe_selection(wm::AbstractWaterModel, n::Int, k::Int, node_fr::Int, node_to::Int, des_pipes::Array{Int,1})
     z_des_pipe = var(wm, n, :z_des_pipe)
     c = JuMP.@constraint(wm.model, sum(z_des_pipe[a] for a in des_pipes) == 1.0)
     append!(con(wm, n, :des_pipe_selection)[k], [c])
@@ -134,7 +134,7 @@ function constraint_on_off_pump_power_custom(wm::AbstractWaterModel, n::Int, a::
 end
 
 
-function constraint_on_off_pump_group(wm::AbstractWaterModel, n::Int, k::Int, pump_indices::Set{Int64})
+function constraint_on_off_pump_group(wm::AbstractWaterModel, n::Int, k::Int, pump_indices::Set{Int})
     pump_indices_sorted = sort(collect(pump_indices))
 
     for i in 1:length(pump_indices_sorted[1:end-1])
@@ -147,14 +147,14 @@ function constraint_on_off_pump_group(wm::AbstractWaterModel, n::Int, k::Int, pu
 end
 
 
-function constraint_on_off_pump_switch(wm::AbstractWaterModel, a::Int, network_ids::Array{Int64, 1}, max_switches::Int64)
+function constraint_on_off_pump_switch(wm::AbstractWaterModel, a::Int, network_ids::Array{Int, 1}, max_switches::Int)
     z_switch_on_sum = sum(var(wm, n, :z_switch_on_pump, a) for n in network_ids)
     c = JuMP.@constraint(wm.model, z_switch_on_sum <= max_switches)
     append!(con(wm, network_ids[end], :on_off_pump_switch)[a], [c])
 end
 
 
-function constraint_pump_switch_on(wm::AbstractWaterModel, a::Int, n_1::Int, n_2::Int, nws_active::Array{Int64, 1})
+function constraint_pump_switch_on(wm::AbstractWaterModel, a::Int, n_1::Int, n_2::Int, nws_active::Array{Int, 1})
     z_1, z_2 = var(wm, n_1, :z_pump, a), var(wm, n_2, :z_pump, a)
     z_switch_on = var(wm, n_2, :z_switch_on_pump, a)
     c_1 = JuMP.@constraint(wm.model, z_switch_on >= z_2 - z_1)
@@ -168,7 +168,7 @@ function constraint_pump_switch_on(wm::AbstractWaterModel, a::Int, n_1::Int, n_2
  end
 
 
- function constraint_pump_switch_off(wm::AbstractWaterModel, a::Int, n_1::Int, n_2::Int, nws_inactive::Array{Int64, 1})
+ function constraint_pump_switch_off(wm::AbstractWaterModel, a::Int, n_1::Int, n_2::Int, nws_inactive::Array{Int, 1})
     z_1, z_2 = var(wm, n_1, :z_pump, a), var(wm, n_2, :z_pump, a)
     z_switch_off = var(wm, n_2, :z_switch_off_pump, a)
     c_1 = JuMP.@constraint(wm.model, z_switch_off >= z_1 - z_2)
